# modules/shipment/openapi.yaml
openapi: 3.1.0
info:
  title: Shipment Module API
  version: 0.1.0
  description: >-
    Task-based HTTP API of the Shipment bounded context. Endpoints are named
    after business tasks (not CRUD) and answer with hypermedia controls
    pointing at the next allowed task.
  contact:
    name: Trade Logistics Management
servers:
  - url: http://localhost:8000
    description: Local development

# Remaining task-based endpoints, added slice by slice:
#   POST /shipments/{id}/flag-exception    -> flag operational exception
paths:
  /shipments:
    post:
      operationId: createDraftShipment
      summary: Create draft shipment
      description: >-
        Opens a shipment in `Draft` state and assigns it a unique waybill
        number drawn from the carrier series. No integration event is
        published: this is a local setup step.
      tags: [Shipment]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDraftShipmentCommand'
      responses:
        '201':
          description: The draft shipment was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDraftShipmentResponse'
        '422':
          description: >-
            The payload is malformed, or the route violates a domain invariant
            (for example the origin equals the destination).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'


  /shipments/{shipment_id}/route:
    put:
      operationId: assignRoute
      summary: Assign complex route
      description: >-
        Replaces the itinerary of a shipment. `transit_legs` lists the
        intermediate calls only, in order; together with the endpoints they
        must form a logical sequence that never calls at the same port twice.
        The route is frozen once the shipment awaits customs release, is in
        transit or has been delivered. No integration event is published.
      tags: [Shipment]
      parameters:
        - name: shipment_id
          in: path
          required: true
          description: Identifier of the shipment to reroute.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRouteCommand'
      responses:
        '200':
          description: The route was assigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignRouteResponse'
        '404':
          description: No shipment matches the identifier.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: >-
            The shipment has advanced past the point where the route can be
            changed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: >-
            The payload is malformed, or the route violates a domain invariant
            (equal endpoints, or a port called at twice).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'


  /shipments/{shipment_id}/finalize-manifest:
    post:
      operationId: finalizeManifest
      summary: Finalize cargo manifest
      description: >-
        Declares the cargo of a draft shipment and moves it to
        `ReadyForManifest`. The total weight must be greater than zero. A
        `ShipmentManifestFinalized` event is written to
        `shipment.outbox_messages` in the same transaction as the state change,
        and published later by the `dispatch-messages` worker.
      tags: [Shipment]
      parameters:
        - name: shipment_id
          in: path
          required: true
          description: Identifier of the shipment to declare.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinalizeManifestCommand'
      responses:
        '200':
          description: The manifest was finalized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinalizeManifestResponse'
        '404':
          description: No shipment matches the identifier.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: The shipment is no longer a draft.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: >-
            The payload is malformed, or the manifest violates a domain
            invariant (a weight of zero or less, a negative volume, or a
            missing commodity code).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

components:
  schemas:
    CreateDraftShipmentCommand:
      type: object
      additionalProperties: false
      required: [origin_port_code, destination_port_code]
      properties:
        origin_port_code:
          type: string
          description: UN/LOCODE of the port where the cargo is loaded.
          pattern: '^[A-Z]{2}[A-Z0-9]{3}$'
          examples: ['ESVLC']
        destination_port_code:
          type: string
          description: UN/LOCODE of the port where the cargo is delivered.
          pattern: '^[A-Z]{2}[A-Z0-9]{3}$'
          examples: ['USNYC']

    CreateDraftShipmentResponse:
      type: object
      required: [shipment_id, waybill_number, status]
      properties:
        shipment_id:
          type: string
          format: uuid
        waybill_number:
          type: string
          pattern: '^MUST-[0-9]{7}$'
        status:
          $ref: '#/components/schemas/TrackingStatus'
        _links:
          type: array
          description: Tasks that can be performed on the shipment next.
          items:
            $ref: '#/components/schemas/Link'
          examples:
            - - rel: assign-route
                href: /shipments/{id}/route
                method: PUT
              - rel: finalize-manifest
                href: /shipments/{id}/finalize-manifest
                method: POST

    AssignRouteCommand:
      type: object
      additionalProperties: false
      required: [origin_port_code, destination_port_code]
      properties:
        origin_port_code:
          type: string
          description: UN/LOCODE of the port where the cargo is loaded.
          pattern: '^[A-Z]{2}[A-Z0-9]{3}$'
          examples: ['ESVLC']
        destination_port_code:
          type: string
          description: UN/LOCODE of the port where the cargo is delivered.
          pattern: '^[A-Z]{2}[A-Z0-9]{3}$'
          examples: ['USNYC']
        transit_legs:
          type: array
          default: []
          description: >-
            Ordered intermediate calls between the origin and the destination.
            The endpoints are not repeated here.
          items:
            type: string
            pattern: '^[A-Z]{2}[A-Z0-9]{3}$'
          examples:
            - ['MAMIR', 'PTLIS']

    AssignRouteResponse:
      type: object
      required: [shipment_id, waybill_number, status, route]
      properties:
        shipment_id:
          type: string
          format: uuid
        waybill_number:
          type: string
          pattern: '^MUST-[0-9]{7}$'
        status:
          $ref: '#/components/schemas/TrackingStatus'
        route:
          $ref: '#/components/schemas/Route'
        _links:
          type: array
          description: Tasks that can be performed on the shipment next.
          items:
            $ref: '#/components/schemas/Link'

    Route:
      type: object
      required: [origin_port_code, destination_port_code, transit_legs]
      properties:
        origin_port_code:
          type: string
        destination_port_code:
          type: string
        transit_legs:
          type: array
          items:
            type: string

    FinalizeManifestCommand:
      type: object
      additionalProperties: false
      required: [total_weight_kg, total_volume_cbm, commodity_code]
      properties:
        total_weight_kg:
          type: number
          description: Gross weight of the cargo, in kilograms.
          examples: [12500.5]
        total_volume_cbm:
          type: number
          description: Gross volume of the cargo, in cubic metres.
          examples: [68.25]
        commodity_code:
          type: string
          maxLength: 16
          description: Customs classification of the goods (HS code).
          examples: ['8471']

    FinalizeManifestResponse:
      type: object
      required: [shipment_id, waybill_number, status, manifest]
      properties:
        shipment_id:
          type: string
          format: uuid
        waybill_number:
          type: string
          pattern: '^MUST-[0-9]{7}$'
        status:
          $ref: '#/components/schemas/TrackingStatus'
        manifest:
          $ref: '#/components/schemas/Manifest'
        _links:
          type: array
          description: Tasks that can be performed on the shipment next.
          items:
            $ref: '#/components/schemas/Link'

    Manifest:
      type: object
      required: [total_weight_kg, total_volume_cbm, commodity_code]
      properties:
        total_weight_kg:
          type: number
        total_volume_cbm:
          type: number
        commodity_code:
          type: string

    Link:
      type: object
      required: [rel, href, method]
      properties:
        rel:
          type: string
        href:
          type: string
        method:
          type: string

    TrackingStatus:
      type: string
      enum:
        - Draft
        - ReadyForManifest
        - AwaitingCustomsRelease
        - InTransit
        - Delivered
        - ExceptionHeld

    ProblemDetails:
      type: object
      description: RFC 9457 error representation returned by every endpoint.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string